+++ test/trending_repo_test.dart
@@ -19,8 +19,10 @@ void main() {
   });
 
   test('non-existent programming language', () async {
-    var result = await ghTrendingRepositories(programmingLanguage: 'dwodjow');
-    expect(result.isEmpty, true);
+    const lang = 'dwodjow';
+    var result = await ghTrendingRepositories(programmingLanguage: lang);
+    final matchingLanguage = result.where((x) => x.programmingLanguage.toLowerCase() == lang);
+    expect(matchingLanguage.isEmpty, true);
   });
 
   test('html only', () async {
@@ -31,8 +33,7 @@ void main() {
   });
 
   test('monthly', () async {
-    var result =
-        await ghTrendingRepositories(dateRange: GhTrendDateRange.thisMonth);
+    var result = await ghTrendingRepositories(dateRange: GhTrendDateRange.thisMonth);
     expect(result.isNotEmpty, true);
     for (var item in result) {
       _checkItem(item, dateRangeMustBe: 'stars this month');
@@ -40,8 +41,7 @@ void main() {
   });
 
   test('weekly', () async {
-    var result =
-        await ghTrendingRepositories(dateRange: GhTrendDateRange.thisWeek);
+    var result = await ghTrendingRepositories(dateRange: GhTrendDateRange.thisWeek);
     expect(result.isNotEmpty, true);
     for (var item in result) {
       _checkItem(item, dateRangeMustBe: 'stars this week');
@@ -49,14 +49,10 @@ void main() {
   });
 
   test('weekly javascript', () async {
-    var result = await ghTrendingRepositories(
-        dateRange: GhTrendDateRange.thisWeek,
-        programmingLanguage: 'javascript');
+    var result = await ghTrendingRepositories(dateRange: GhTrendDateRange.thisWeek, programmingLanguage: 'javascript');
     expect(result.isNotEmpty, true);
     for (var item in result) {
-      _checkItem(item,
-          programmingLanguageMustBe: 'javascript',
-          dateRangeMustBe: 'stars this week');
+      _checkItem(item, programmingLanguageMustBe: 'javascript', dateRangeMustBe: 'stars this week');
     }
   });
 }
@@ -70,11 +66,9 @@ void _checkItem(
     expect(item.owner.isNotEmpty, true);
     expect(item.repoName.isNotEmpty, true);
     if (programmingLanguageMustBe.isNotEmpty) {
-      expect(item.programmingLanguage.toLowerCase(),
-          programmingLanguageMustBe.toLowerCase());
+      expect(item.programmingLanguage.toLowerCase(), programmingLanguageMustBe.toLowerCase());
     }
-    if (item.programmingLanguage.isNotEmpty &&
-        item.programmingLanguageColor != null) {
+    if (item.programmingLanguage.isNotEmpty && item.programmingLanguageColor != null) {
       expect(item.programmingLanguageColor!.isNotEmpty, true);
     }
     expect(item.starsSince.isNotEmpty, true);