+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.3
+
+- [@leoafarias](https://github.com/leoafarias) fixed the issue where the fork count is the total star count. [Pull-request](https://github.com/xamantra/gh_trend/pull/1)
+- Fixed test for non-existent programming language in the trending repositories.
+
 ## 1.0.2
 
 - Added programming language color.